+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -4,9 +4,6 @@ import blackjack.domain.Dealer;
 import blackjack.domain.Player;
 import blackjack.domain.Players;
 import blackjack.dto.ResultDto;
-import blackjack.manager.BlackJackInitManager;
-import blackjack.manager.BlackJackResultManager;
-import blackjack.manager.BlackjackProcessManager;
 import blackjack.view.Confirmation;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
